[Refactor] Organize LTX DiffVAE under ltx2/vae
Move the existing Diffusers decoder adapter and distributed wrapper into the VAE package, updating imports without changing execution.
