+++ vllm_omni/diffusion/models/ltx2/vae/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM-Omni project
+
+"""Ordinary LTX-2.5 DiffVAE, tiled decoding and acceleration operators."""
